fix output softmax visualization
For the "SM Agg" matrix at the bottom of nano-gpt, the labels were
flipped from the values. That is, the row containing the max values was
labeled exp, and the row containing the exp values was labeled max.

The cause was that in this odin model, those rows were flipped compared
to their counterparts in src/llm/GptModelLayout.ts and
src/llm/GptModel.ts. Flipping them back fixes the problem.

Also updates the path to native.wasm in build.sh. (build.sh got moved
down a directory in 0ba3a39 "Move llm code into its own folder", but
this relative path wasn't updated.)

Also deletes native.wasm.o. It's not clear to me that this was
intentionally committed in the first place. build.sh deleted it
automatically.